refactor: 입력값 검증 InputView 에서 하도록 변경
